backend/requirements/base.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9

# Task Queue & Cache
celery==5.3.6
redis==5.0.1

# AI & Crawling
crawl4ai==0.3.74
anthropic==0.18.1
openai==1.12.0

# LangChain (for agent orchestration)
langchain==0.1.9
langchain-anthropic==0.1.4
langchain-openai==0.0.5

# Utilities
python-dotenv==1.0.0
pydantic==2.6.1
PyYAML==6.0.1
requests==2.31.0

# Development helpers
httpx==0.26.0